Move the WebSocket client into streamedlogs.py
